+++ vulkano-shaders/src/parse.rs
@@ -9,43 +9,8 @@
 
 use enums::*;
 
-/// Parses a SPIR-V document.
-pub fn parse_spirv(data: &[u8]) -> Result<Spirv, ParseError> {
-    if data.len() < 20 {
-        return Err(ParseError::MissingHeader);
-    }
-
-    // we need to determine whether we are in big endian order or little endian order depending
-    // on the magic number at the start of the file
-    let data = if data[0] == 0x07 && data[1] == 0x23 && data[2] == 0x02 && data[3] == 0x03 {
-        // big endian
-        data.chunks(4)
-            .map(|c| {
-                     ((c[0] as u32) << 24) | ((c[1] as u32) << 16) | ((c[2] as u32) << 8) |
-                         c[3] as u32
-                 })
-            .collect::<Vec<_>>()
-
-    } else if data[3] == 0x07 && data[2] == 0x23 && data[1] == 0x02 && data[0] == 0x03 {
-        // little endian
-        data.chunks(4)
-            .map(|c| {
-                     ((c[3] as u32) << 24) | ((c[2] as u32) << 16) | ((c[1] as u32) << 8) |
-                         c[0] as u32
-                 })
-            .collect::<Vec<_>>()
-
-    } else {
-        return Err(ParseError::MissingHeader);
-    };
-
-    parse_u32s(&data)
-}
-
-/// Parses a SPIR-V document from a list of u32s.
-///
-/// Endianness has already been handled.
-fn parse_u32s(i: &[u32]) -> Result<Spirv, ParseError> {
+/// Parses a SPIR-V document from a list of words.
+pub fn parse_spirv(i: &[u32]) -> Result<Spirv, ParseError> {
     if i.len() < 5 {
         return Err(ParseError::MissingHeader);
     }
@@ -399,6 +364,12 @@ mod test {
     #[test]
     fn test() {
         let data = include_bytes!("../tests/frag.spv");
-        parse::parse_spirv(data).unwrap();
+        let insts: Vec<_> = data.chunks(4)
+            .map(|c| {
+                ((c[3] as u32) << 24) | ((c[2] as u32) << 16) | ((c[1] as u32) << 8) | c[0] as u32
+            })
+            .collect();
+
+        parse::parse_spirv(&insts).unwrap();
     }
 }